定时任务执行批量更新时出现空指针异常问题解决方法

+++ liwei-framework/src/main/java/com/liwei/utils/RedisCache.java
@@ -202,7 +202,7 @@ public class RedisCache
     }
 
 
-    public void incrementCacheMapValue(String key,String hKey,int v){
+    public void incrementCacheMapValue(String key,String hKey,long v){
         redisTemplate.opsForHash().increment(key,hKey,v);
     }
